Refactor components folder
