+++ src/components/commons/Button.tsx
@@ -1,4 +1,4 @@
-import { ButtonProps } from "../types/types";
+import { ButtonProps } from "../../types/types";
 
 const Button: React.FC<ButtonProps> = ({
   type,
@@ -9,7 +9,6 @@ const Button: React.FC<ButtonProps> = ({
   danger,
   disabled,
 }) => {
-  
   const buttonClasses =
     "flex justify-center rounded-md px-3 py-2 text-sm font-semibold focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2" +
     (disabled ? " opacity-50 cursor-default" : "") +
